Execute archive/delete actions only once per label.
When archiving, there is no copy happening; when deleting, it's not
necessary to remove the label (actually, it can be worse, as then the
label is no longer visible in the Trash folder on the webmail).

Since only one of the two action has to be executed, it's possible to
apply a batch operation on the messages, archiving or deleting them at
once. It was not possible when doing copy and flag setting, as then
the second operation would get a different input set from the first.